Updated Staging view to show more traked fares
Previously we only displayed fares for today and the day before, or from the expiration date.
This means if a person forwarded an itineray many days after purchasing ticket, it could be difficult to find appropriate fare.
Now it displays cached fares for every day after the fare protection was purchased.